Imagej-ROI-filter-centroids2svg/requirements.txt
roifile
shapely
svgwrite
pillow
